blackeagle/serializers/projectsSerializer.py
from rest_framework import serializers
from ..models.projects import Projects


class ProjectsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Projects
        fields = '__all__'
