requirements.txt
anyio==4.4.0
asgiref==3.7.2
certifi==2024.7.4
click==8.1.7
dj-database-url==2.1.0
Django==4.2.9
django-admin-list-filter-dropdown==1.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-json-widget==2.0.1
djangorestframework==3.3.2
exceptiongroup==1.2.2
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-telegram-bot==21.4
pytz==2023.3.post1
sniffio==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
uvicorn==0.27.0.post1
whitenoise==6.7.0
